requirements.txt
pydantic
PyYAML
python-dotenv
requests
psycopg2-binary
websockets
